+++ hellodagger/src/androidTest/java/com/airbnb/mvrx/hellodagger/Extensions.kt
@@ -1,4 +1,4 @@
-package com.airbnb.mvrx.helloDagger2
+package com.airbnb.mvrx.hellodagger
 
 import androidx.annotation.IdRes
 import androidx.fragment.app.Fragment
