+++ hellodagger/src/main/java/com/airbnb/mvrx/hellodagger/base/BaseViewModel.kt
@@ -1,8 +1,8 @@
-package com.airbnb.mvrx.helloDagger2.base
+package com.airbnb.mvrx.hellodagger.base
 
 import com.airbnb.mvrx.BaseMvRxViewModel
 import com.airbnb.mvrx.MvRxState
-import com.airbnb.mvrx.helloDagger2.BuildConfig
+import com.airbnb.mvrx.hellodagger.BuildConfig
 
 /**
  * Base class for ViewModels.
