+++ hellodagger/src/main/java/com/airbnb/mvrx/hellodagger/di/AppComponent.kt
@@ -1,6 +1,6 @@
-package com.airbnb.mvrx.helloDagger2.di
+package com.airbnb.mvrx.hellodagger.di
 
-import com.airbnb.mvrx.helloDagger2.base.BaseViewModel
+import com.airbnb.mvrx.hellodagger.base.BaseViewModel
 import dagger.Component
 
 @Component(modules = [AppModule::class])
